Revert "#design: 캘린더 페이지 구현"
