--- src/pages/Diary/Diary.jsx
@@ -1,9 +1,0 @@
-import Navigation from '../../components/Common/Navbar'
-
-export default function Diary(){
-  return (
-    <>
-      <Navigation />
-    </>
-  )
-}
