--- src/pages/Todo/TodoList.jsx
@@ -1,9 +1,0 @@
-import Navigation from '../../components/Common/Navbar'
-
-export default function TodoList(){
-  return (
-    <>
-      <Navigation />
-    </>
-  )
-}
